backend/requirements.txt
asgiref==3.11.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
Django==6.0.1
django-appconf==1.2.0
django-cors-headers==4.9.0
django-cryptography==1.1
django-environ==0.12.0
django-filter==25.2
django-jazzmin==3.0.1
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.11
pillow==12.1.0
psycopg==3.3.2
psycopg2-binary==2.9.11
pycparser==2.23
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis==7.1.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
stripe==14.2.0
typing_extensions==4.15.0
urllib3==2.6.3
